+++ EfSamuariDomain/Entities/Moves/Block.cs
@@ -4,7 +4,7 @@ namespace EfSamuariDomain.Entities.Moves
 {
     public class Block : IDefensiveMove
     {
-        public int DamageReduction { get; set; }
+        public float DamageReduction { get; set; }
 
     }
 }
